chore: sweeping rename of Core API components (#3896)
Part 1: shorten names and use a consistent part of speech:

    core_context.training -> core_context.train
    core_context.checkpointing -> core_context.checkpoint
    core_context.preemption -> core_context.preempt
    # core_context.distributed stays because it matches e.g. torch.distributed
    # core_context.searcher stays because we like it better.

Part 2: Change the class name of each component of the Core API to
include "Context".  This won't really affect users but it's a bit clearer.

    core.Searcher -> core.SearcherContext
    core.Training -> core.TrainContext
    core.Preemption -> core.PreemptContext
    core.Checkpointing -> core.CheckpointContext
    # core.DistributedContext remains unchanged.

Part 3: avoid abbreviations

    core_context.search.ops -> core_context.search.operations
    core.SearcherOp -> core.SearcherOperation

Part 4: anything that must only be called by the chief is named report_*

    SearcherOperation.complete() -> SearcherOperation.report_completed()

Part 5: expose DistributedContext._zmq_*, stripping the _zmq_ prefix.
Originally, the prefix was meant to make it clear that
context.distributed.allgather() was definitely not hvd.allgather(), but
we care enough about not marrying the name to a particular communication
library that we are willing to sacrifice the clear hvd/zmq distinction.

    DistributedContext._zmq_allgather() -> DistributedContext.allgather()
    DistributedContext._zmq_gather() -> DistributedContext.gather()
    DistributedContext._zmq_broadcast() -> DistributedContext.broadcast()
    # etc